Fixed data types and filter functions from  redundant encounter slot loads so it doesn't try and load pointers a billion fucking trillion times and failing randomly because I'm bad at C. Fuck!

+++ include/seeds.h
@@ -9,7 +9,7 @@
 
 typedef struct {
     uint32_t timer;
-    uint16_t seed;
+    uint32_t seed;
 } InitialSeed;
 
 typedef struct {
